requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
imageio==2.34.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
pillow==10.4.0
pip==23.2.1
requests==2.32.3
setuptools==68.2.0
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.3
wheel==0.41.2
